Add initial testing
